requirements.txt
Flask
Flask-Cors
requests
Gunicorn
python-dotenv